+++ NEWS.md
@@ -4,10 +4,12 @@
 
 * Add new function `to_lm()` which converts `ardl`, `uecm`, and `recm` models, 
 as well as `dynlm` models in `lm` models. Among other things, this helps using 
-the model with other functions (e.g. with `predict()` to forecast).
+the model with other functions (e.g. with `predict()` to forecast)
+(issue #5, @melville1808).
 
 * `multipliers()` now has a new argument (`se`) which allows the estimation of
 standard errors for the delay multipliers, using the delta method.
+(issue #17, @sammy-w)
 
 * Add new function `plot_delay()` which creates plots (`ggplot`) for the delay 
 multipliers and their uncertainty intervals (standard errors or confidence